refactor(test): FilterMaster
- Check unity among jaxb marshaller and XmlMapper
